Sistema puntos de recarga, puntuaciones, etc casi completo Usuarios 	Hecho 		Botón eliminar cuenta			-- HECHO 		Cambio de email	y password		-- HECHO 		Pop up para todos los cambios 	-- HECHO 	TODO 		Clase Vehíclo 					-- De momento descartado, solo si hay tiempo 		Propiedad autonomía Mapa 	TODO 		Mapa principal 			- Añadir Mapa 			- Quizás limitar la funcionalidad para que solo se vean los puntos de recarga y las paradas que marque el usuario (origen destino) 			- Añadir marcadores 			- Calcular distancia puntos Puntos recarga 	TODO 		- Necesario que firebase escuche por actualziaciones en la DB de puntos de recarga e informe de esto para actualizar los datos 			--> En su defecto, que cada vez que se accede a una vista en la que hay puntos de recarga que se vuelvan a descargar los datos (actualmente esto) 		- Comprobar si punto de recarga esta repetido 				De momento solo compruebo si la latitud y longitud estan entre 0.995 y 1.005 de los de otros puntos de recarga 		- El texto de "Verificado" no esta puesto ni vincualdo en todos lados, revisar
	Clase Punto recarga
		HECHO
			Subclase Parada 		------------- HECHO
			Añadir					------------- HECHO
			Propiedades 			------------- HECHO
		TODO
			Eliminar 				Habrá que añadir un buscador (La primera categoria, en vez de ser ruta, que sea lista completa, y que tenga un campo de búsqueda)
										Si eres el creador del punto de recarga que aparezca un icono "Papelera" en vez de el boton de nuevo
			Modificar 				Que el boton de modificar solo aparezca si ers el usuario que lo ha creado
										Si no eres el usuario que lo ha creado que aparezca un boton de reportar que abre un cuadro de diálogo para que pngas un motivo y listo

	Info punto
		Subsistema Puntuaciones
			HECHO
				Añadir puntuación   	------------- HECHO
				Ver puntuación			------------- HECHO
				Modificar puntuación	------------- HECHO
				Propiedades 			------------- HECHO
			TODO
				Eliminar puntuación		Añadir boton en la parte inferior derecha

+++ TFG1/app/src/main/java/es/usal/tfg1/model/Puntuacion.java
@@ -3,7 +3,7 @@ package es.usal.tfg1.model;
 public class Puntuacion {
     private String id;
     private String comentario;
-    private double puntuacion;
+    private float puntuacion;
 
     public String getId() {
         return id;
@@ -26,22 +26,22 @@ public class Puntuacion {
         }
     }
 
-    public double getPuntuacion() {
-        return puntuacion;
+    public float getPuntuacion() {
+        return (float)puntuacion;
     }
 
-    public void setPuntuacion(double puntuacion) {
+    public void setPuntuacion(float puntuacion) {
         this.puntuacion = puntuacion;
     }
 
     public Puntuacion() { }
 
-    public Puntuacion(String id, String comentario, double puntuacion) {
+    public Puntuacion(String id, String comentario, float puntuacion) {
         this.id = id;
         this.comentario = comentario;
         this.puntuacion = puntuacion;
     }
-    public Puntuacion(String id, double puntuacion) {
+    public Puntuacion(String id, float puntuacion) {
         this.id = id;
         this.puntuacion = puntuacion;
         this.comentario = "";
@@ -52,4 +52,5 @@ public class Puntuacion {
         this.puntuacion = p.puntuacion;
         this.comentario = p.comentario;
     }
+
 }
